IOS-6198 Extract widget SubscriptionData builder and harden dataView wait
