+++ Anytype/Sources/PresentationLayer/Modules/HomeWidgets/Container/Models/SetSubscriptionDataBuilder+Widget.swift
@@ -1,0 +1,23 @@
+import Foundation
+import Services
+
+extension SetSubscriptionDataBuilderProtocol {
+    func widgetSubscriptionData(
+        widgetInfo: BlockWidgetInfo,
+        setDocument: any SetDocumentProtocol,
+        identifier: String,
+        spaceType: SpaceType?
+    ) -> SubscriptionData {
+        let setSubData = SetSubscriptionData(
+            identifier: identifier,
+            document: setDocument,
+            groupFilter: nil,
+            currentPage: 0,
+            numberOfRowsPerPage: widgetInfo.fixedLimit,
+            collectionId: setDocument.isCollection() ? setDocument.objectId : nil,
+            objectOrderIds: setDocument.objectOrderIds(for: subscriptionId),
+            spaceType: spaceType
+        )
+        return set(setSubData)
+    }
+}
